refactor: [docs] Remove Adsterra ads
It is found that Adsterra ads are malware and had negative impact on
site performance. Site viewers complained multiple times and this is
documented in blogs easily found by google search. So it makes sense to
remove this now and use another ad network.

Another change done is to remove ads styles from theme styles as this is
not part of the theme so should not be defined here instead use a static
style CSS file.